--- lib/features/about/presentation/presentation.dart
@@ -1,5 +1,0 @@
-// Pages.
-export 'pages/about_page.dart';
-
-// Providers.
-export 'providers/package_info_provider.dart';
